Issue #985 Add thank you message
